% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{check_parks}
\alias{check_parks}
\title{check_parks}
\usage{
check_parks(path.root = NULL)
}
\arguments{
\item{path.root}{location of current dropbox root Example: ~/Dropbox/pkg.data/}
}
\description{
checks parks for raw and clean shapes
}
\keyword{denver}
\keyword{parks_check}
