% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_fetching.R
\name{clean_geometry}
\alias{clean_geometry}
\title{Clean geometries safely in projected space}
\usage{
clean_geometry(x)
}
\arguments{
\item{x}{An sf object}
}
\value{
A cleaned sf object
}
\description{
Projects to EPSG:3857, repairs, then transforms back to EPSG:4326
}
