% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_address.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode an address using tidygeocoder}
\usage{
geocode_address(address)
}
\arguments{
\item{address}{A character string of a single address.}
}
\value{
A tibble with latitude and longitude columns.
}
\description{
This function takes a street address and returns latitude and longitude using the Nominatim API.
}
\examples{
geocode_address("10 Downing Street, London")
}
