% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocode_coords}
\alias{geocode_coords}
\title{Convert coordinates into a geocoded point}
\usage{
geocode_coords(lat, lon)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}
}
\value{
A tibble with lat and lon columns.
}
\description{
This function wraps a user-input pair of latitude and longitude into a named tibble.
}
\examples{
location <- geocode_coords(49.793, 9.932)
}
