% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_fetching.R
\name{get_roads}
\alias{get_roads}
\title{Get road network within a travel zone}
\usage{
get_roads(zone)
}
\arguments{
\item{zone}{An \code{sf} polygon from \code{travel_zone()}}
}
\value{
An \code{sf} object of road lines or NULL if none found
}
\description{
Returns roads (highways) that intersect a travel zone.
}
