% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{parks_check}
\alias{parks_check}
\title{parks_check}
\usage{
parks_check(path.root = NULL)
}
\arguments{
\item{path.root}{location of current dropbox root Example: ~/Dropbox/pkg.data/}
}
\description{
checks parks for raw and clean shapes
}
\keyword{denver}
\keyword{parks_check}
