% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{parks_get}
\alias{parks_get}
\title{parks_get}
\usage{
parks_get(path.root = NULL, proj.name = NULL, fresh = FALSE)
}
\arguments{
\item{path.root}{location of current dropbox root Example: ~/Dropbox/pkg.data/}

\item{proj.name}{Projection to use for project default: denver}

\item{fresh}{Redo import and clean raw shapes}
}
\description{
load and cleans parks shapefiles
}
\keyword{denver}
\keyword{google}
\keyword{load}
\keyword{panoids}
\keyword{parks}
