% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate_to_target.R
\name{route_length}
\alias{route_length}
\title{Calculate and save the length of a route}
\usage{
route_length(route)
}
\arguments{
\item{route}{An sf LINESTRING object representing a route, from "navigate_to_target()"}
}
\value{
The length of the route in a numeric unit (meters)
}
\description{
Calculate and save the length of a route
}
