% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode an address using tidygeocoder}
\usage{
geocode_address(address)
}
\arguments{
\item{address}{A character string of a single address. Special characters such as Umlauts or "ß" should work.}
}
\value{
A tibble with latitude and longitude columns.
}
\description{
This function takes a street address and returns latitude and longitude using the Nominatim API.
}
\examples{
geocode_address("John-Skilton-Straße 4a, 97074 Würzburg")
}
